Refactor(.remind) : 리마인드 정보 생성 클래스 의존관계 변경
