+++ src/test/java/com/newbarams/ajaja/module/remind/application/SchedulingRemindServiceTest.java
@@ -16,7 +16,6 @@ import com.newbarams.ajaja.module.plan.domain.Message;
 import com.newbarams.ajaja.module.plan.domain.Plan;
 import com.newbarams.ajaja.module.plan.domain.RemindInfo;
 import com.newbarams.ajaja.module.remind.application.model.RemindMessageInfo;
-import com.newbarams.ajaja.module.remind.application.port.in.CreateRemindUseCase;
 import com.newbarams.ajaja.module.remind.application.port.out.FindRemindablePlanPort;
 import com.newbarams.ajaja.module.remind.application.port.out.SendPlanInfoRemindPort;
 
@@ -29,7 +28,7 @@ class SchedulingRemindServiceTest extends MockTestSupport {
 	@Mock
 	private SendPlanInfoRemindPort sendPlanInfoRemindPort;
 	@Mock
-	private CreateRemindUseCase createRemindUseCase;
+	private CreateRemindService createRemindService;
 
 	@Mock
 	private Plan plan;
